fix: randomNumber
I was trying to call math.random() and then tried to call its result with the argument of (0*4). However, it returns a float between 0 and 1, it can't be called.
Moreover 0 times 4 equals 0, so it was redundant.

The correct way is to utilize math.random() to generate said float between 0 and 1, and then multiply it by 4, so it ranges from 0 to 4 (inclusive)