fix: animatePress
I attempted to use the delay() method directly with addClass() and removeClass() methods. However, the delay() method in jQuery works only with animations in the animation queue, such as fadeIn(), fadeOut(), slideUp(), etc. To fix this issue, I decided to use setTimeout() function to delay the removal of the pressed class after a certain period.